backend/requirements.txt
flask>=3.1.1
flask-cors>=6.0.1
flask-limiter>=4.0.0
apscheduler>=3.11.1
pillow>=12.0.0
numpy>=2.2.6
requests>=2.32.5
paddlepaddle>=3.2.2
paddleocr>=3.3.2
PyJWT>=2.10.1
python-dotenv>=1.2.1
xlsxwriter>=3.2.5
reportlab>=4.4.5